refactor(core): split the deserializer by concern, dispatch styles by table
`deserialize-parameter.ts` had grown to 1129 lines mixing three unrelated
jobs. Split along the seams that were already there:

  deserialize/schema.ts     what type is this, really — through $ref,
                            allOf, anyOf/oneOf, enum, const — and is this
                            wire string an unambiguous lexical form of it.
                            Knows nothing about styles.
  deserialize/style/*.ts    one module per serialization style, reversing
                            the wire form into scalar strings. Knows
                            nothing about types beyond "is this a list".
  deserialize/items.ts      the shared typing step both feed.
  deserialize/result.ts     the result union and its two messages.
  deserialize/split.ts      delimiter splitting, including RFC 9110 lists.

`deserialize-parameter.ts` is now 45 lines of public surface, so every
existing import keeps working.

Path styles are dispatched through a table rather than a chain of `if`s.
Which styles a location supports is part of that location's shape, and a
reader should not have to reconstruct it from conditionals — the review
noted that the unsupported branch had become unreachable for any
spec-valid path document without that being visible anywhere.

Shape is unchanged: same 488 tests, and the reference document produces
byte-identical violations.

Refs #624
